Set river level__m as float

+++ src/flood_forecaster/data_model/river_level.py
@@ -27,13 +27,11 @@ class PredictedRiverLevel(Base):
     id = Column(Integer, primary_key=True)
     location_name = Column(String(100))
     date = Column(DateTime)
-    level_m = Column(Integer)
+    level_m = Column(Float)
     station_number = Column(String(50))
     ml_model_name = Column(String(100))
     forecast_days = Column(Integer, comment="Number of days into the future the forecast is for")
     risk_level = Column(String(50), comment="Risk level of the forecasted river level, e.g., 'Low', 'Medium', 'High'")
-
-# TODO: add orm getters and setters
 
 
 class StationDataFrameSchema(pa.DataFrameModel):
